Update Date.bi
game playing to discover how to create a branch which seems to be specifically tied to adding files to the repo and being prompted for new branch.  or perhaps i made that up. but look how directly below this box is a radio button choice to commit to this branch or create an even new branch. so yes i think this must be how it is done